True Interval Patch, second part
The best interval dialog rounded intervals to the nearest second
due to a casting of a double to int. This was introduced by Mark L
during the intervals code patch and is an error.

All the plots have now been adjusted to correctly determine if a ride
point is within an interval. Related cropping and binning issues in
3d plot an Histogram plot have also been corrected.

fixes #15

+++ src/BestIntervalDialog.cpp
@@ -229,12 +229,12 @@ BestIntervalDialog::findClicked()
         resultsTable->setItem(row, 2, n);
 
         // hidden columns - start, stop
-        QString strt = QString("%1").arg((int) interval.start); // can't use secs as it gets modified
+        QString strt = QString("%1").arg(interval.start); // can't use secs as it gets modified
         QTableWidgetItem *st = new QTableWidgetItem;
         st->setText(strt);
         resultsTable->setItem(row, 3, st);
 
-        QString stp = QString("%1").arg((int)(interval.start + x));
+        QString stp = QString("%1").arg(interval.start + x);
         QTableWidgetItem *sp = new QTableWidgetItem;
         sp->setText(stp);
         resultsTable->setItem(row, 4, sp);
